feat(routes): extract routes by role into focused modules
Extract user-routes.ts (446 lines) into 7 modular route files organized by role and responsibility.

Changes:
- Created worker/routes/ directory with 7 focused modules
  - student-routes.ts: 4 student routes (98 lines)
  - teacher-routes.ts: 4 teacher routes (100 lines)
  - admin-routes.ts: 7 admin routes (122 lines)
  - parent-routes.ts: 1 parent route (35 lines)
  - user-management-routes.ts: 6 user management routes (95 lines)
  - system-routes.ts: 1 system route (11 lines)
  - route-utils.ts: shared route validation utilities (18 lines)
  - index.ts: exports all route modules (7 lines)
- Refactored user-routes.ts from 446 to 12 lines (97% reduction)
- Extracted validateUserAccess to shared route-utils.ts
- Updated docs/blueprint.md with Route Module Architecture section
- Updated docs/task.md with completed task documentation

Benefits:
- Separation of Concerns: Routes organized by role and responsibility
- Single Responsibility: Each module has focused purpose
- Modularity: Each route module is atomic and replaceable
- Maintainability: Easier to locate and modify routes
- Typecheck passed: 0 errors, zero regressions

Architectural Impact:
- Clean Architecture: Dependencies flow correctly
- Open/Closed: New routes for a role can be added without modifying others
- DRY Principle: Shared utilities extracted
