+++ docs/task.md
@@ -9401,6 +9401,151 @@ const mockStudentService = createStudentService(new MockRepository());
 
 None currently in progress.
 
+---
+
+### Route Module Extraction (2026-01-09) - Completed ✅
+
+**Task**: Extract routes by role into focused modules following Separation of Concerns and Single Responsibility principles
+
+**Problem**:
+- worker/user-routes.ts had 446 lines with routes for all user roles mixed in one file
+- Violated Separation of Concerns: Routes for different roles were intermixed
+- Violated Single Responsibility Principle: File had multiple reasons to change (student routes, teacher routes, admin routes, etc.)
+- Difficult to maintain: Finding routes for a specific role required searching through large file
+- Not modular: Could not replace or test routes for one role independently
+
+**Solution**:
+- Extracted routes into 7 focused modules organized by role and responsibility
+- Created worker/routes/ directory with separate files for each route category
+- Extracted shared utilities to route-utils.ts
+- Refactored user-routes.ts to import and register all route modules
+- Achieved clean separation of concerns and modularity
+
+**Implementation**:
+
+1. **Created worker/routes/ directory** with 7 new modules:
+   - **student-routes.ts** (98 lines): 4 student routes (grades, schedule, card, dashboard)
+   - **teacher-routes.ts** (100 lines): 4 teacher routes (dashboard, announcements, grades, announcements)
+   - **admin-routes.ts** (122 lines): 7 admin routes (dashboard, users, announcements, settings, rebuild-indexes)
+   - **parent-routes.ts** (35 lines): 1 parent route (dashboard)
+   - **user-management-routes.ts** (95 lines): 6 user management routes (CRUD for users and grades)
+   - **system-routes.ts** (11 lines): 1 system route (database seeding)
+   - **route-utils.ts** (18 lines): Shared route validation and helper functions
+   - **index.ts** (7 lines): Exports all route modules and utilities
+
+2. **Refactored user-routes.ts** from 446 lines to 12 lines:
+   - Removed all 23 route handlers (lines 50-445)
+   - Removed validateUserAccess function (extracted to route-utils.ts)
+   - Added imports for all route modules from worker/routes
+   - Created registry function that registers all route modules
+   - Maintained backward compatibility: Same function signature and export
+
+3. **Extracted Shared Utilities** to route-utils.ts:
+   - validateUserAccess() function for user access validation
+   - Reusable across all route modules
+   - Centralized logging and error handling for access denied
+
+**Metrics**:
+
+| Metric | Before | After | Improvement |
+|---------|---------|--------|-------------|
+| user-routes.ts lines | 446 | 12 | 97% reduction |
+| Total route modules | 0 | 7 | New modular structure |
+| Route module lines | 0 | 486 | Distributed across modules |
+| Average module size | N/A | 69 lines | Focused, maintainable |
+| Separation of Concerns | Mixed | Clean | Complete separation |
+| Single Responsibility | Multiple concerns | Focused modules | All principles met |
+| Typecheck errors | 0 | 0 | No regressions |
+| Cognitive load | High (search 446 lines) | Low (focused modules) | Significantly reduced |
+
+**Benefits Achieved**:
+- ✅ user-routes.ts reduced by 97% (446 → 12 lines)
+- ✅ Routes organized by role and responsibility
+- ✅ Each route module is atomic and replaceable
+- ✅ Separation of Concerns achieved (routes separated by role)
+- ✅ Single Responsibility achieved (each module has one reason to change)
+- ✅ Shared utilities extracted to route-utils.ts
+- ✅ Easier to locate routes (all student routes in one file)
+- ✅ Reduced cognitive load (each file is focused and maintainable)
+- ✅ New routes for a role can be added to that role's module without modifying others
+- ✅ Typecheck passed with 0 errors (no regressions)
+- ✅ Zero breaking changes to existing functionality
+
+**Technical Details**:
+
+**Route Module Organization**:
+- Each module exports a function that registers routes on Hono app
+- Modules import necessary dependencies (services, entities, middleware)
+- Functions are pure (no side effects except route registration)
+- Type-safe Context parameter for better IDE support
+
+**Registration Pattern**:
+```typescript
+// worker/user-routes.ts (registry)
+import { studentRoutes, teacherRoutes, adminRoutes, ... } from './routes';
+
+export function userRoutes(app: Hono<{ Bindings: Env }>) {
+  studentRoutes(app);      // Register all student routes
+  teacherRoutes(app);     // Register all teacher routes
+  adminRoutes(app);       // Register all admin routes
+  parentRoutes(app);      // Register all parent routes
+  userManagementRoutes(app);  // Register all user management routes
+  systemRoutes(app);       // Register all system routes
+}
+```
+
+**Shared Utilities**:
+- `validateUserAccess()`: Validates that authenticated user can access requested resource
+- Parameters: Context, userId, requestedId, role, resourceType
+- Returns boolean: true if access allowed, false if denied
+- Logs access denied warnings with contextual information
+- Calls forbidden() response for denied access
+- Imported from worker/type-guards: getCurrentUserId() for getting authenticated user ID
+
+**Architectural Impact**:
+- **Modularity**: Each route module is atomic and replaceable
+- **Separation of Concerns**: Routes organized by role and responsibility
+- **Clean Architecture**: Dependencies flow correctly (user-routes → route modules → services → entities)
+- **Single Responsibility**: Each module handles one concern (role's routes)
+- **Open/Closed**: New routes for a role can be added without modifying others
+- **DRY Principle**: Shared utilities extracted to route-utils.ts (no duplication)
+- **Maintainability**: Easier to understand, locate, and modify routes
+
+**Success Criteria**:
+- [x] worker/routes/ directory created with 7 route modules
+- [x] student-routes.ts extracted (4 routes, 98 lines)
+- [x] teacher-routes.ts extracted (4 routes, 100 lines)
+- [x] admin-routes.ts extracted (7 routes, 122 lines)
+- [x] parent-routes.ts extracted (1 route, 35 lines)
+- [x] user-management-routes.ts extracted (6 routes, 95 lines)
+- [x] system-routes.ts extracted (1 route, 11 lines)
+- [x] route-utils.ts created with shared utilities
+- [x] user-routes.ts refactored to 12 lines (97% reduction)
+- [x] Separation of Concerns achieved (routes organized by role)
+- [x] Single Responsibility achieved (each module focused)
+- [x] Typecheck passed with 0 errors
+- [x] Zero breaking changes to existing functionality
+
+**Impact**:
+- `worker/routes/`: New directory with 7 route modules (486 total lines)
+  - student-routes.ts: 98 lines (4 student routes)
+  - teacher-routes.ts: 100 lines (4 teacher routes)
+  - admin-routes.ts: 122 lines (7 admin routes)
+  - parent-routes.ts: 35 lines (1 parent route)
+  - user-management-routes.ts: 95 lines (6 user management routes)
+  - system-routes.ts: 11 lines (1 system route)
+  - route-utils.ts: 18 lines (shared utilities)
+  - index.ts: 7 lines (exports)
+- `worker/user-routes.ts`: Refactored from 446 → 12 lines (97% reduction)
+- `docs/blueprint.md`: Added Route Module Architecture section with diagrams and metrics
+- Modularity: 7 focused modules instead of 1 monolithic file
+- Maintainability: Each route module is independently testable and replaceable
+- Separation of Concerns: Complete separation by role and responsibility
+
+**Success**: ✅ **ROUTE MODULE EXTRACTION COMPLETE, ROUTES ORGANIZED BY ROLE, USER-ROUTES.TS REDUCED BY 97%**
+
+---
+
 ### CircuitBreaker Module Extraction (2026-01-09) - Completed ✅
 
 **Task**: Extract CircuitBreaker from api-client.ts to separate resilience module
