+++ worker/routes/index.ts
@@ -1,0 +1,7 @@
+export { studentRoutes } from './student-routes';
+export { teacherRoutes } from './teacher-routes';
+export { adminRoutes } from './admin-routes';
+export { userRoutes as userManagementRoutes } from './user-management-routes';
+export { parentRoutes } from './parent-routes';
+export { systemRoutes } from './system-routes';
+export { validateUserAccess } from './route-utils';
